Rename .hpp files to .h for consistency.
