+++ src/profilerlib/Cli.h
@@ -13,7 +13,7 @@
 #include <string>
 #include <vector>
 
-#include "Tui.hpp"
+#include "Tui.h"
 
 /*
  * CLI to parse command line arguments and start up the TUI to either run a single command,
